Added functionality to Render comments dynamically
